--- src/test/java/com/cryptomorin/xseries/Test1_16R1.java
@@ -1,15 +1,0 @@
-package com.cryptomorin.xseries;
-
-import com.cryptomorin.xseries.versions.Spigot_1_16R1;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-final class Test1_16R1 extends Spigot_1_16R1 {
-
-    @Test
-    void testXMaterial() {
-        runServer(() -> {
-            Assertions.assertTrue(XMaterial.matchXMaterial("AIR").isPresent());
-        });
-    }
-}
