+++ src/test/java/com/cryptomorin/xseries/TestUtil.java
@@ -1,0 +1,17 @@
+package com.cryptomorin.xseries;
+
+import org.junit.jupiter.api.Assertions;
+
+public final class TestUtil {
+
+    public static final Runnable USUAL_TEST = () -> {
+        testXMaterial();
+    };
+
+    private TestUtil() {
+    }
+
+    public static void testXMaterial() {
+        Assertions.assertTrue(XMaterial.matchXMaterial("AIR").isPresent());
+    }
+}
