+++ src/test/java/com/cryptomorin/xseries/v1_16R2/Test1_16R2.java
@@ -1,0 +1,12 @@
+package com.cryptomorin.xseries.v1_16R2;
+
+import com.cryptomorin.xseries.TestUtil;
+import org.junit.jupiter.api.Test;
+
+final class Test1_16R2 extends Spigot_1_16R2 {
+
+    @Test
+    void test() {
+        runServer(TestUtil.USUAL_TEST);
+    }
+}
