+++ src/test/java/com/cryptomorin/xseries/v1_16R3/Test1_16R3.java
@@ -1,0 +1,12 @@
+package com.cryptomorin.xseries.v1_16R3;
+
+import com.cryptomorin.xseries.TestUtil;
+import org.junit.jupiter.api.Test;
+
+final class Test1_16R3 extends Spigot_1_16R3 {
+
+    @Test
+    void test() {
+        runServer(TestUtil.USUAL_TEST);
+    }
+}
